refactor(mainpage): 수업관리 화면과 학생관리 화면에 헤더 추가
