--- src/pages/ManageClass/main.module.scss
@@ -1,17 +1,0 @@
-@use 'src/styles/constants/colors' as color;
-
-.manageAttendancePage {
-  width: 100%;
-  height: 100%;
-  overflow: scroll;
-  background-color: color.$WHITE;
-
-  .pageTitle {
-    margin-top: 50px;
-    margin-bottom: 25px;
-    margin-left: 50px;
-    font-size: 26px;
-    font-weight: 700;
-    color: color.$PRIMARY;
-  }
-}
