+++ src/pages/ManageStudent/manageStudent.module.scss
@@ -6,16 +6,10 @@
   overflow: scroll;
   background-color: color.$WHITE;
 
-  .pageTitle {
-    margin-top: 50px;
-    margin-bottom: 25px;
-    margin-left: 50px;
-    font-size: 26px;
-    font-weight: 700;
-    color: color.$PRIMARY;
-  }
-
   .chartWrap {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
     padding: 35px;
   }
 }
